Updates per review on #71
